requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
face-recognition>=1.3.0
numpy>=1.24.0
psycopg2-binary>=2.9.6
websockets>=11.0.3
gradio>=3.50.0
pillow>=10.0.0
opencv-python>=4.8.0.74
python-dotenv>=1.0.0
requests>=2.31.0